requirements.txt
django
djangorestframework
environs
drf-spectacular
